requirements.txt
Flask==0.10.1
Flask-Cors==1.9.0
# Celery[mongodb]==3.1.16
https://github.com/totem/celery/archive/patch_task-deserialize-url-fix.tar.gz
flower==0.7.2
pymongo==2.7.2
gevent==1.0.1
uWSGI==2.0.7
supervisor==3.1.2
python-etcd==0.3.2
elasticsearch==1.2.0
https://github.com/totem/fleet-py/archive/master.tar.gz
https://github.com/totem/yoda-py/archive/master.tar.gz
repoze.lru==0.6
jsonschema==2.4.0
ordered-set==1.3